refactor: Move icon component to its new home.
