+++ package.json
@@ -15,7 +15,7 @@
     "db:deploy": "prisma migrate deploy --schema ./db/schema.prisma",
     "db:studio": "prisma studio --schema ./db/schema.prisma",
     "icons:add": "sly add",
-    "icons:build": "icons build -i ./icons -o ./app/components/icon/icons --optimize 0",
+    "icons:build": "icons build -i ./icons -o ./app/components/ui/icon/icons --optimize 0",
     "check": "biome check --apply .",
     "typecheck": "tsc"
   },
